Fix remaining bugs and clean up unused code
- Fix HotelRepository ID type: Integer → Long to match Hotel.id field
- Fix RoleRepository ID type: Long → Integer to match Role.id field
- Cascade Hotel ID type fix: int → long in HotelService, HotelsServiceImpl, HotelController
- Remove unused HotelRepository injection from HotelController
- Fix @UniqueConstraint column name: "username" → "user_name" to match Spring naming strategy
- Fix jwtExpiration type: int → long in JwtUtils
- Fix parameter name typo: authentification → authentication in JwtUtils
- Fix log message typo: autentification → authentication in AuthTokenFilter
- Remove unused imports: Page/Pageable (HotelService), coyote.Response/SecurityContext (AuthController), StringUtil (AuthTokenFilter)

+++ src/main/java/com/hotel/is/dao/HotelRepository.java
@@ -3,5 +3,5 @@ package com.hotel.is.dao;
 import com.hotel.is.entity.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel, Integer> {
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
 }
